+++ lib/src/components/components.dart
@@ -27,6 +27,7 @@ export 'slider/foss_slider.dart';
 export 'spinner/foss_spinner.dart';
 export 'switch/foss_switch.dart';
 export 'tabs/foss_tabs.dart';
+export 'text/foss_text.dart';
 export 'text_field/foss_text_field.dart';
 export 'toast/foss_toast.dart';
 export 'toast/foss_toast_controller.dart';